Refactor reference table codec and update CRC helper
